Refactor libSQL Viewer integration: remove unused components, streamline iframe source logic, and enhance URL handling
