--- packages/studiocms_devapps/src/routes/libsql-viewer.astro
@@ -1,26 +1,0 @@
----
-import { dbEnv } from 'virtual:studiocms-devapps/config';
-import { getIFrameSrc } from '../utils/libsql-url.js';
----
-<!DOCTYPE html>
-<html lang="en">
-<head>
-  <meta charset="UTF-8">
-  <meta name="viewport" content="width=device-width, initial-scale=1.0">
-  <title>libSQL Database Viewer</title>
-</head>
-<body style="background: black;">
-  <iframe 
-    id="sqlIframe" 
-    title="libSQL Database Viewer"
-    src={getIFrameSrc(dbEnv.remoteUrl)}
-    data-url={dbEnv.remoteUrl} 
-    data-authtoken={dbEnv.token}
-    style="height: 94vh; width: 98vw; border: 0;"
-    ></iframe>
-
-  <script>
-    import "virtual:studiocms-devapps/libsql/createClient";
-  </script>
-</body>
-</html>
